Saved. Slot 1

+++ 0x00-hello_world/101-quote.c
@@ -9,7 +9,7 @@ int main(void)
 {
 	int i = 0;
 	char quote[] = "and that piece of art is useful\" - Dora Korpar, 2015-10-19";
-	while (i < sizeof(quote))
+	while (i < (int) sizeof(quote))
 	{
 		putchar(quote[i]);
 		i++;
